refactor(RepositoryTest): RepositoryTest 코드 패키지 변경
