+++ src/test/java/gift/integration/repository/JpaUserRepositoryTest.java
@@ -1,4 +1,4 @@
-package gift.repository.integration;
+package gift.integration.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
