--- src/test/java/gift/repository/unit/JpaProductRepositoryTest.java
@@ -1,4 +1,0 @@
-package gift.repository.unit;
-
-class JpaProductRepositoryTest {
-}
