--- src/test/java/gift/repository/unit/JpaWishRepositoryTest.java
@@ -1,4 +1,0 @@
-package gift.repository.unit;
-
-class JpaWishRepositoryTest {
-}
